Separate "QuestionList" out from the "question_time_setting" page and revise the according test
